board crud api

+++ community/backend/src/middleware/controller/board.controller.ts
@@ -2,8 +2,9 @@ import { Context } from 'koa'
 import { ResponseDto } from '../repository/dto/response.dto';
 import { BoardEntity, SearchBoardDto } from '../repository/dto/board.dto';
 import { setEntityParameters } from '../utils/common.util';
-import { RESULT_CODE } from '../../types';
+import { RESULT_CODE, STATE_YN } from '../../types';
 import * as boardService from '../service/board.service'
+import { CustomError } from '../utils/custom.error';
 
 /**
  * 글 조회
@@ -53,6 +54,13 @@ export const updateBoard = async (ctx: Context) => {
     const reqParam = new BoardEntity();
     setEntityParameters(reqParam, ctx.request.body);
 
+    if(!reqParam.boardIdx){
+        throw new CustomError(
+            RESULT_CODE.INVALID_PARAMETER.code,
+            RESULT_CODE.INVALID_PARAMETER.msg
+        )
+    }
+
     const resultData = await boardService.updateBoard(reqParam);
 
     result.data = resultData;
@@ -69,11 +77,19 @@ export const updateBoard = async (ctx: Context) => {
 export const deleteBoard = async (ctx: Context) => {
   const result = new ResponseDto();
   try {
-    const reqParam = new SearchBoardDto();
+    const reqParam = new BoardEntity();
     setEntityParameters(reqParam, ctx.request.body);
 
-    const resultData = await boardService.deleteBoard(reqParam);
+    if(!reqParam.boardIdx){
+        throw new CustomError(
+            RESULT_CODE.INVALID_PARAMETER.code,
+            RESULT_CODE.INVALID_PARAMETER.msg
+        )
+    }
 
+    reqParam.delYn = STATE_YN.Y
+    
+    const resultData = await boardService.deleteBoard(reqParam);
     result.data = resultData;
     result.setResultCode(RESULT_CODE.SUCCESS);
   } catch (e) {
